+++ blog/templates/blog/widgets/list_card.html
@@ -11,9 +11,9 @@
     <!-- card-bodyの左側 -->
     <div class="col-xs-12 col-sm-6 text-center">
       {% if post.thumnail %}
-        <a href="{{ post.thumnail.url }}" target="_blank" rel="nofollow"><img class="img-fluid lazy" data-original="{{ post.thumnail.url }}"></a>
+        <a href="{{ post.thumnail.url }}" target="_blank" rel="nofollow"><img class="img-fluid lazy" data-original="{{ post.thumnail.url }}" alt="{{ post.title }}"></a>
       {% else %}
-        <img class="img-fluid lazy" data-original="{% static 'blog/img/noimage.png' %}">
+        <img class="img-fluid lazy" data-original="{% static 'blog/img/noimage.png' %}" alt="{{ post.title }}">
       {% endif %}
     </div><!-- card-bodyの左側終わり -->
